pointcloud_octomap_updater changed to remove clipped cells
Clipped cells marked as free in this change.
This was not the case before which caused the clipped cells to stay resulting in many undesired wrongly occupied voxels. This is especially the case for a dynamic scene observed by the depth camera.

The voxels now update correctly with no clipped cells at unoccupied spaces.