Update .hadolint.yaml
I am Naman Sharma and In this updated version of the code, I've added a new ignored rule "SC1001". This rule is used by the Shellcheck linter to detect issues with shell scripts, such as using shell features that are not supported by all shells. By ignoring this rule, the linter tool will no longer report this type of issue.

The effect of this change is that the linter tool will now ignore the "SC1001" rule when checking code, which can be useful if you have shell scripts that use features that are not supported by all shells, but which are still necessary for your application to function correctly. However, it's important to note that ignoring rules can sometimes lead to missed errors or security vulnerabilities, so it's important to use this feature with caution and only after carefully considering the potential risks.