Update package.xml
I am Naman Sharma and I have made the following changes The modifications made to the original code are as follows:

Added comments to describe the purpose of different elements in the file. This improves readability and understanding.

Ensured consistent formatting throughout the file, such as indentation and line breaks. This makes the code easier to read and understand.

Used more descriptive names for elements where possible. This improves readability and understanding.

Removed commented-out code to make the file more concise and easier to read.

These modifications should not have any significant effect on the functionality of the package, but they can make the code easier to read and maintain, which is important for the long-term development and evolution of the software.