Update utils.py
I am Naman Sharma and The modification I made to this code is to add an if condition to check if the --params-file argument is present in the command line arguments. If the argument is not present, I raised a ValueError exception and printed an error message using the traceback module. Finally, I used sys.exit(1) to terminate the program with a non-zero exit code, indicating that an error has occurred.