Update trajopt_planner_manager.cpp
I am Naman Sharma Added #include <ros/ros.h>: This header file was missing in the original code, and is required to use ROS_INFO and ros::NodeHandle.

Added #include <moveit_msgs/MoveItErrorCodes.h>: This header file is required to use moveit_msgs::MoveItErrorCodes.

Changed ROS_ERROR to ROS_WARN in the getPlanningContext() function: The original code used ROS_ERROR to print a warning message, which would be printed in red in the terminal. I changed it to ROS_WARN, which prints the message in yellow and is more appropriate for a warning.

Added a check to getPlanningContext() to ensure that the requested group name is valid: The original code assumed that the requested group name is always valid, but this may not always be the case. I added a check to ensure that the requested group name exists in the planning_contexts_ map before attempting to retrieve the corresponding TrajOptPlanningContext.

Added a check to getPlanningContext() to ensure that the TrajOptPlanningContext object for the requested group name has been initialized: The original code assumed that the TrajOptPlanningContext object for the requested group name has been initialized during initialize(), but this may not always be the case. I added a check to ensure that the object exists before attempting to use it.

The effects of these changes are as follows:

Adding #include <ros/ros.h> allows the code to use ROS_INFO and ros::NodeHandle.

Adding #include <moveit_msgs/MoveItErrorCodes.h> allows the code to use moveit_msgs::MoveItErrorCodes.

Changing ROS_ERROR to ROS_WARN ensures that the warning message is printed in yellow instead of red.

Adding a check to getPlanningContext() to ensure that the requested group name is valid prevents the code from attempting to use an uninitialized TrajOptPlanningContext.

Adding a check to getPlanningContext() to ensure that the TrajOptPlanningContext object for the requested group name has been initialized prevents the code from attempting to use a null PlanningContextPtr.