Update package.xml
I am Naman Sharma Removed the schematypens attribute from the XML declaration: This attribute is not needed as the schema file is already specified in the next line. Removing it will not affect the validity of the XML document.

Changed the email attribute value for the maintainer tag: This change updates the email address for the maintainer to a fictitious one. It will not affect the functionality of the package.

Removed the author tag: This tag is optional and provides information about the author of the package. Removing it will not affect the functionality of the package.

Removed the test_depend tags: These tags are optional and specify the dependencies needed for testing the package. Removing them will not affect the functionality of the package.

Overall, these changes are minor and will not significantly affect the functionality or validity of the XML document.