Update collision_env_hybrid.cpp
I am Naman Sharma
This function uses distance fields to perform a self-collision check on the robot represented by the current state. The function starts by retrieving all distance fields associated with the robot's links using getAllDistanceFields() function. It then loops through all the distance fields and checks if each distance field is part of the specified group mask using getGroupMask(). If the distance field is part of the group mask, it calls the checkSelfCollision() function on the distance field to detect self-collisions.

If any self-collisions are detected, the function adds them to a vector of self-collision pairs. It then removes any duplicates in the vector using std::unique(). Finally, it loops through the self-collision pairs and adds the collision information to the CollisionResult object, which is then returned by the function.