Adding planning_frame and ee_frame to panda_simulated_config.yaml
When implementing moveit servo for my own robot. The servo node crashed because planning_frame and ee_frame were not defined. This was a bit of a pain to debug because the error did not show up in the terminal log when the servo node was started as a node component (this may also be fixed).

Disclaimer: I have not had time to test if this works with the example provided. So either wait until I do, or test it on your machine before approving.