Exit earlier on failure in `generatePlan` (#2726)
With this change, the `PlanningPipeline::generatePlan()` exits as soon
as a failure is detected. Before this, `break` was used to exit the
current loop of request adapters, planners, or response adapters, but
the function continued to the next loop. For example, if a planner would
fail, the response adapters would still be executed.

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>
Co-authored-by: Gaël Écorchard <gael@km-robotics.cz>