moveit_ros/perception/pointcloud_octomap_updater: log missing parameters
When the `sensor_3d.yaml` file with `PointCloudUpdater` is used, but
some required parameters are missing, `moveit_group` node fails to start
with the error log: `Failed to configure updater of type
PointCloudUpdater`, which doesn't give any clues of what's actually
failed.

This PR is aimed to provide more user-friednly log message on telling
explicitly of what is missing, e.g.:
```
[move_group-2] [ERROR] [1750954248.813778881] [move_group]: Missing parameters under 'vision_perception_node': , 'max_update_rate', 'filtered_cloud_topic'
[move_group-2] [ERROR] [1750954248.814011991] [move_group.moveit.moveit.ros.occupancy_map_monitor]: Failed to configure updater of type PointCloudUpdater
```

The changes is done only in the `pointcloud_octomap_updater.cpp`, but if
accepted it can be also re-used in `depth_image_octomap_updater.cpp`
which has the same behaviour.